\name{gaussnoise}
\alias{gaussnoise}
\docType{data}
\title{Gaussian Clusters with White Noise}
\description{
Simulated toy data set to demonstrate the MCGFA package. Generated using the \code{mnormt} package for R.
}
\usage{data(gaussnoise)}
\format{
  A data frame with 205 observations and 6 columns. The first column gives the class number: \code{1} and \code{2} are the two Gaussian clusters. The second column indicates whether an observation arose as noise. (This is eqivalent to the first column being equal to \code{3}.)
}
\keyword{datasets}
