\name{mcgfa-package}
\alias{mcgfa-package}
\docType{package}
\title{
    mcgfa: Model-Based Clustering and Classification with Mixtures of Contaminated Gaussian Factor Analyzers
}
\description{
    Performs clustering and classification using the Mixtures of Contaminated Gaussian Factor Analyzers model. Allows for automatic detection of outliers and noise.

}
\details{
    \tabular{ll}{
        Package: \tab mcgfa\cr
        Type: \tab Package\cr
        Version: \tab 1.0\cr
        Date: \tab 2015-06-13\cr
        License: \tab GPL (>=2)\cr
        LazyLoad: \tab yes\cr
    }
}
\author{
    Martin Blostein, Antonio Punzo, and Paul D. McNicholas

    Maintained by: Martin Blostein <martin.blostein@gmail.com>
}
\references{
    ...
}
\keyword{ package }
\seealso{
    \code{\link{mcgfa}} for main function, and details.
}
